Task bullet space added
`tasks_bullet_space` allows setting the bullet and margin space type.
This makes it possible to fully use of tabs instead of spaces. For
instance, The following configuration will work nicely in PlainTasks:
```json
{
	"tasks_bullet_space": "\t",
	"translate_tabs_to_spaces": false
}
```
Having a Tab as a separator allow the `done`, `cancelled` and `pending`
tasks to always align without "jumping" around when the state changes.