Don't use "horrible" color/font/font-size schemes by default.
If they are not set by default, global sublime settings are used, which is
good for some users (me). This settings can't be changed in user settings
to be ignored (so the global settings would be used), you can only change
them to something else, but never to the global defaults.

I am really tired of editing PlainTasks default file, which keep changing
after each update back.