Implement commands for adding due date and for sorting by due date and priority
A command for adding due date is useful in Sublime Text 4 as the completion menu offers quite a few choices and the snippet for inserting the due date is not the first choice in the menu (when doing d+TAB). The command instead is one key away all the time. Adding due date is bound to F4 key by default. It supports multiple cursors too.

Sorting by due date and priority works by first sorting by due date (and leaving the items without a due date at the end and secondly sorting by priority with the following order: @critical, @high, normal (doesn't exist so any task without other priority tags is considered normal priority), and lastly @low. Sorting is done only on the first level allowing fine control on what to sort, nested lists are supported and just the list to be sorted can be picked. To sort you have to place the cursor on the header of target list. Sorting is bound to F5 by default. It supports multiple cursors too (being able to sort multiple lists at once).

Reverse sorting is also supported and is bound by default to F7.